Massively simplify the parity-complex solver.
The previous check actually built a state and called validate(), but there were
only two ways it could fail: sequence rule on this exact row, or uniqueness
rule. Dealing with the uniqueness rule isn't really the job of this simple
solver, so we can ignore that; and there's no way to get a triple-1 without
having first tripped one of the other checks (.xx or xx.), so the only way it
could have failed validation is if it resulted in "000". So just check for that!